Refactor: clean up app router
