+++ src/app/(after-login)/(nav)/question/_components/searched-questions.tsx
